filedialog
Because of the structure of tkinter, you need to explicitly import the filedialog submodule. (see this link: https://stackoverflow.com/questions/36163561/why-tkinter-module-raises-attribute-error-when-run-via-command-line-but-not-when?utm_medium=organic&utm_source=google_rich_qa&utm_campaign=google_rich_qa).  Could you change this slightly so that the program initializes with self.chemfile set to the therm.cti file in the folder containing the code? That way, the user can just use the default chemistry file and they don't have to select a new chemistry file each time they open the program.